Fixed cloud backup purge
The D argument in {rclone --config $rclone_conf  delete  --min-age d $rclone_retention $rclone_disk:"$rclone_path"} is not a valid use of the --min-age flag. in order to use the the --min-age flag argument must be a number and a letter indicating it as day, week, month, or year. EX. 10d, 10w, 10m, 10y. These values can be stored at "rclone_retention="